Finsihed algorithm for finding dates.
