+++ dates.txt
@@ -1,3 +1,4 @@
+.2010 - wife; nightmares and angry outbursts; drinking a lot; living in deceased grandmother's house with wife.  Working for American Electric Power- billing.
 03/25/93 Total time of visit (in minutes):
 6/18/85 Primary Care Doctor:
 sshe plans to move as of 7/8/71 In-Home Services: None
