+++ text_mining_package/__init__.py
@@ -1,1 +1,2 @@
-from nic_date_date_class import NicDate
+from text_mining_package.date_finder import DateFinder
+from text_mining_package.nic_date_date_class import NicDate
